+++ src/typescript/macros/weaponAttack.ts
@@ -26,16 +26,20 @@ export default async function assignWeaponAttackMacro(
 async function getOrCreateWeaponAttackMacro(
   data: WeaponAttackDragData
 ): Promise<Macro | undefined> {
-  const actor = getGame().actors?.get(data.actorId);
-  if (!(actor instanceof Actor)) return;
+  const actor = data.actorId ? getGame().actors?.get(data.actorId) : null;
 
-  const weapon = actor.items.get(data.weaponId);
+  let weapon;
+  if (actor) {
+    weapon = actor.items.get(data.weaponId);
+  } else {
+    weapon = getGame().items?.get(data.weaponId);
+  }
   if (!(weapon instanceof Weapon)) return;
 
   const name = createMacroName(
-    actor.toObject(),
     weapon.toObject(),
-    data.attackName
+    data.attackName,
+    actor?.toObject()
   );
   const command = createMacroCommand(data);
 
@@ -51,18 +55,19 @@ async function getOrCreateWeaponAttackMacro(
 }
 
 /**
- * Create a Macro name from the Actor, Weapon data and Attack name.
- * @param actorData - the source data of the Actor, the Macro is for
+ * Create a Macro name from the optional Actor, Weapon data and Attack name.
  * @param weaponData - the source data of the Weapon, the Macro is for
  * @param attackName - the name of the Attack, the Macro is for
+ * @param actorData - the source data of the Actor, the Macro is for
  * @returns the Macro name
  */
 function createMacroName(
-  actorData: foundry.data.ActorData["_source"],
   weaponData: foundry.data.ItemData["_source"],
-  attackName: string
+  attackName: string,
+  actorData?: foundry.data.ActorData["_source"] | null | undefined
 ): string {
-  return `${actorData.name}/${weaponData.name}/${attackName}`;
+  const tail = `${weaponData.name}/${attackName}`;
+  return actorData ? `${actorData.name}/${tail}` : tail;
 }
 
 /**
@@ -71,24 +76,33 @@ function createMacroName(
  * @returns the Macro command
  */
 function createMacroCommand(data: WeaponAttackDragData): string {
-  return `game.wv.macros.executeWeaponAttack("${data.actorId}", "${data.weaponId}", "${data.attackName}")`;
+  const args = [data.weaponId, data.attackName];
+  if (data.actorId) args.push(data.actorId);
+
+  return `game.wv.macros.executeWeaponAttack("${args.join('","')}")`;
 }
 
 /**
- * Execute a Weapon Attack Macro on an Actor.
- * @param actorId - the ID of the Actor, the Weapon belongs to
- * @param weaponId - the ID of the Weapon on the Actor
+ * Execute a Weapon Attack Macro.
+ * @param weaponId - the ID of the Weapon
  * @param attackName - the name of the Attack on the Weapon
+ * @param actorId - the ID of the Actor, the Weapon belongs to, blank if unowned
  */
 export function executeWeaponAttack(
-  actorId: string,
   weaponId: string,
-  attackName: string
+  attackName: string,
+  actorId?: string | null | undefined
 ): void {
-  const actor = getGame().actors?.get(actorId);
-  if (!(actor instanceof Actor)) return;
+  let weapon;
 
-  const weapon = actor.items.get(weaponId);
+  if (typeof actorId === "string" && actorId !== "") {
+    const actor = getGame().actors?.get(actorId);
+    if (!(actor instanceof Actor)) return;
+
+    weapon = actor.items.get(weaponId);
+  } else {
+    weapon = getGame().items?.get(weaponId);
+  }
   if (!(weapon instanceof Weapon)) return;
 
   const attack = weapon.systemData.attacks.attacks[attackName];